refactor(move): 修改CashDelta和相关函数的字段类型，优化数据结构

+++ move/tycoon/sources/events.move
@@ -111,7 +111,7 @@ public struct CashDelta has copy, drop, store {
     is_debit: bool,
     amount: u64,
     reason: u8,  // 1=toll, 2=buy, 3=upgrade, 4=bonus, 5=fee, 6=card
-    details: u64
+    details: u16
 }
 
 // 卡牌获取项
@@ -313,7 +313,7 @@ public(package) fun make_cash_delta(
     is_debit: bool,
     amount: u64,
     reason: u8,
-    details: u64
+    details: u16
 ): CashDelta {
     CashDelta {
         player,
